# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GophProfile API
  description: Microservice for user avatar management.
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Local server
  - url: http://gophprofile.local
    description: Kubernetes ingress (via client nginx)

tags:
  - name: avatars
  - name: health

paths:
  /api/v1/avatars:
    post:
      tags: [avatars]
      summary: Upload avatar
      security:
        - UserID: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Avatar image
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarUploadResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/avatars/{avatar_id}:
    get:
      tags: [avatars]
      summary: Get avatar by ID
      parameters:
        - name: avatar_id
          in: path
          required: true
          description: Avatar UUID
          schema:
            type: string
            format: uuid
        - name: size
          in: query
          description: Thumbnail size
          schema:
            type: string
            enum: [100x100, 300x300, original]
        - name: format
          in: query
          description: Output format
          schema:
            type: string
            enum: [jpeg, png, webp]
      responses:
        "200":
          description: OK
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /health/ready:
    get:
      tags: [health]
      summary: Readiness check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

components:
  securitySchemes:
    UserID:
      type: apiKey
      in: header
      name: X-User-ID

  schemas:
    AvatarUploadResponse:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        status:
          type: string
        url:
          type: string
        created_at:
          type: string
          format: date-time

    HealthResponse:
      type: object
      additionalProperties:
        type: string
      example:
        database: ok
        minio: ok
        rabbitmq: ok
        worker: ok

    ErrorResponse:
      type: object
      additionalProperties:
        type: string
      example:
        error: bad request